Move styled components to styled files, add ToggleIcons to expandable items in sections, minor refactors and cleanup
