+++ src/components/resume/sections/Volunteer/index.tsx
@@ -1,67 +1,19 @@
-import React, { useState } from 'react';
-import styled from 'styled-components';
+import React, { MouseEventHandler, useState } from 'react';
+import { ExpandLess, ExpandMore } from '@mui/icons-material';
 
-import { GenericHeading, GenericContainer, LocationText, NameAndLocationWrapper } from '../components';
-import { themeColors } from '../../../themes';
-
-const VolunteerItemContainer = styled.div`
-    display: flex;
-    flex-direction: column;
-    height: auto;
-    z-index: 1;
-
-    &.collapsed {
-        flex: 0;
-        max-height: min-content;
-        /* padding: 0.5em 10vw; */
-        padding-top: 0.5em;
-        padding-bottom: 0.5em;
-        transition: all 300ms ease-out;
-    }
-
-    &.expanded {
-        flex: 1;
-        max-height: 100%;
-        /* padding: 1em 10vw; */
-        padding-top: 1em;
-        padding-bottom: 1em;
-        transition: all 300ms ease-in;
-    }
-`;
-
-const OrganizationName = styled.h3`
-    margin-bottom: 0;
-    transition: all 300ms ease-in;
-    z-index: 1;
-
-    &:hover {
-        cursor: pointer;
-    }
-`;
-
-const VolunteerItemDetailWrapper = styled.div`
-    display: flex;
-    flex-direction: column;
-    margin: 0;
-    z-index: 0;
-
-    &.hidden,
-    &.hidden * {
-        color: transparent;
-        flex: 0;
-        height: 0;
-        max-height: 0;
-        opacity: 0;
-    }
-
-    &.visible,
-    &.visible * {
-        flex: 1;
-        height: auto;
-        max-height: 100%;
-        opacity: 1;
-    }
-`;
+import {
+    GenericHeading, // <- to force formatting
+    GenericContainer,
+    LocationText,
+    NameAndLocationWrapper,
+    ToggleIcon,
+} from '../../components';
+import { themeColors } from '../../../../themes';
+import {
+    VolunteerItemContainer, // <- to force formatting
+    OrganizationName,
+    VolunteerItemDetailWrapper,
+} from './styled';
 
 interface VolunteerItemProps {
     volunteerRecord: VolunteerRecord;
@@ -70,12 +22,19 @@ interface VolunteerItemProps {
 const VolunteerItem = ({ volunteerRecord }: VolunteerItemProps): React.ReactElement => {
     const [isCollapsed, setIsCollapsed] = useState(true);
 
+    const handleToggleOnClick = (): void => setIsCollapsed(!isCollapsed);
+
     const { organization, roles } = volunteerRecord;
 
     return (
         <VolunteerItemContainer className={`${isCollapsed ? 'collapsed' : 'expanded'}`}>
             <NameAndLocationWrapper>
-                <OrganizationName onClick={() => setIsCollapsed(!isCollapsed)}>{organization.name}</OrganizationName>
+                <ToggleIcon onClick={handleToggleOnClick as MouseEventHandler}>
+                    {isCollapsed ? <ExpandLess /> : <ExpandMore />}
+                </ToggleIcon>
+                <OrganizationName onClick={handleToggleOnClick as MouseEventHandler}>
+                    {organization.name}
+                </OrganizationName>
                 <LocationText className={`${isCollapsed ? 'hidden' : 'visible'}`}>
                     {organization.location.city}, {organization.location.state}
                 </LocationText>
